+++ packages/browser-wallet/src/popup/pages/AgeProofRequest/i18n/da.ts
@@ -1,0 +1,14 @@
+import type en from './en';
+
+const t: typeof en = {
+    moreDetails: 'Flere detajler',
+    verify: 'Verificer',
+    notAbleToVerify: 'Det er ikke muligt at verificere',
+    ageMin: 'Verificer at du er ældre end {{ age }}',
+    ageMax: 'Verificer at du er yngre end {{ age }}',
+    ageBetween: 'Verificer at du er mellem {{ ageMin }} og {{ ageMax }} år gammel',
+    ageExact: 'Verificer at du er {{ age }} år gammel',
+    failure: 'Verifikation mislykkedes',
+};
+
+export default t;