+++ packages/browser-wallet/src/popup/pages/AgeProofRequest/i18n/en.ts
@@ -1,0 +1,12 @@
+const t = {
+    moreDetails: 'More details',
+    verify: 'Verify',
+    ageMin: 'Verify that you are {{age}} years or older',
+    ageMax: 'Verify that you are {{age}} years or younger',
+    ageBetween: 'Verify that you are {{ageMin}} to {{ageMax}} years old',
+    ageExact: 'Verify that you are {{ age }} years old',
+    notAbleToVerify: 'Unable to verify',
+    failure: 'Verification failed',
+};
+
+export default t;