+++ packages/browser-wallet/src/popup/shell/i18n/locales/en.ts
@@ -24,6 +24,7 @@ import web3IdProofRequest from '@popup/pages/Web3ProofRequest/i18n/en';
 import verifiableCredential from '@popup/pages/VerifiableCredential/i18n/en';
 import addWeb3IdCredential from '@popup/pages/AddWeb3IdCredential/i18n/en';
 import verifiableCredentialBackup from '@popup/pages/VerifiableCredentialBackup/i18n/en';
+import ageProofRequest from '@popup/pages/AgeProofRequest/i18n/en';
 
 const t = {
     shared,
@@ -52,6 +53,7 @@ const t = {
     web3IdProofRequest,
     verifiableCredential,
     verifiableCredentialBackup,
+    ageProofRequest,
 };
 
 export default t;