+++ packages/browser-wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+-   A simple page for web3Id age proofs.
+
+## 1.1.7
+
+### Fixed
+
+-   Changed 'Zero Knowledge' to 'Zero-knowledge' in display texts.
+-   An issue with images in verifiable credentials for lower resolutions.
+-   Schema validation of verifiable credentials with attributes that have `{ type: "integer" }` no longer rejects BigInt values.
+
 ## 1.1.6
 
 ### Fixed